Update scraping engine to match daily use engine
In the months since this was originally adapted from the main.rb script, the main engine has changed somewhat.
This commit updates scrape.rb to include the changes from main.rb, as well as adding byebug for debugging purposes. This is the first step to moving the entire project server-side.